docs plugin: split up plugin.go, remove stale code
