--- html/example-pipeline.tmpl
@@ -1,1 +1,0 @@
-<div class="example-pipeline"><div class="pipeline-grid"><div class="serial-grid"><div class="node input" id="node-7"><a href="#">booklit</a></div><div class="node" id="node-0"><div class="job succeeded"><a class="job-name" href="#">unit</a><a class="job-status" href="#"></a></div></div></div></div></div>
