--- html/literate-segment.tmpl
@@ -1,9 +1,0 @@
-<div class="literate-segment">
-  <div class="literate-entry">
-    <div class="prose">
-      {{.Content | render}}
-    </div>
-
-    {{.Partial "Code" | render}}
-  </div>
-</div>
